auth-server/requirements.txt
Flask
pymongo
mongoengine
pyjwt